Reverting changes to include MostReadSection wrapper styling where it is used.
