+++ src/app/pages/FrontPage/index.jsx
@@ -26,7 +26,6 @@ import { ServiceContext } from '#contexts/ServiceContext';
 import FrontPageSection from '#containers/FrontPageSection';
 import MetadataContainer from '#containers/Metadata';
 import MostReadContainer from '#containers/MostRead';
-import { MostReadSection } from '#containers/MostRead/Canonical';
 import LinkedData from '#containers/LinkedData';
 import ATIAnalytics from '#containers/ATIAnalytics';
 import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
@@ -63,15 +62,6 @@ export const StyledFrontPageDiv = styled.div`
   }
 `;
 
-const FrontPageMostReadSection = styled(MostReadSection)`
-  /* To centre page layout for Group 4+ */
-  margin: 0 auto;
-  width: 100%; /* Needed for IE11 */
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
-    max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN};
-  }
-`;
-
 const FrontPage = ({ pageData, mostReadEndpointOverride }) => {
   const {
     product,
@@ -99,8 +89,8 @@ const FrontPage = ({ pageData, mostReadEndpointOverride }) => {
   const renderMostRead = () => (
     <MostReadContainer
       mostReadEndpointOverride={mostReadEndpointOverride}
-      wrapper={FrontPageMostReadSection}
       maxTwoColumns
+      isOnFrontPage
     />
   );
