+++ src/app/pages/FrontPage/index.test.jsx
@@ -165,7 +165,6 @@ describe('Front Page', () => {
         <FrontPageWithContext pageData={pageData} />,
       );
       const sections = container.querySelectorAll('section');
-      console.log('--------', sections);
 
       expect(sections).toHaveLength(2);
       sections.forEach(section => {
